requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
pydantic
joblib
pmdarima
matplotlib
python-dotenv